pyproject.toml
[tool.poetry]
name = "keda_scaling_demo"
version = "0.1.0"
description = ""
authors = ["Ibrahim Halil Koyuncu"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.5"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
